Healthlink - Copy/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# LangChain 1.x with Google GenAI integration
# Requires Python 3.10+
langchain>=1.0.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-google-genai>=3.0.0

# Vector Store
pinecone-client>=3.0.0

# Local Embeddings (offline upsert script)
sentence-transformers>=3.0.0
transformers>=4.48.0
tokenizers>=0.21,<0.22

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Streamlit UI
streamlit==1.30.0
plotly==5.18.0
gradio>=4.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
 

 # Development (move to dev requirements if needed)
